refactored pulse instruction
