+++ src/qat/ir/instructions.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 from copy import deepcopy
 from pydoc import locate
-from typing import Any, Dict, List, Literal, Optional, Set, Union
+from typing import Any, Dict, List, Literal, Optional, Union
 
-import numpy as np
 from compiler_config.config import InlineResultsProcessing
 from pydantic import (
     BaseModel,
@@ -16,16 +15,10 @@ from pydantic import (
 )
 from typing_extensions import Annotated
 
-from qat.purr.compiler.devices import PulseChannel, PulseShapeType, QuantumComponent, Qubit
+from qat.purr.compiler.devices import PulseChannel, QuantumComponent, Qubit
 
 # The following things from legacy instructions are unchanged, so just import for now.
-from qat.purr.compiler.instructions import (
-    AcquireMode,
-    PostProcessType,
-    ProcessAxis,
-    build_generated_name,
-)
-from qat.purr.utils.logger import get_default_logger
+from qat.purr.compiler.instructions import build_generated_name
 
 
 ### Sweeps
@@ -483,251 +476,6 @@ class PhaseReset(GroupInstruction):
         return f"phase reset {','.join(self.targets)}"
 
 
-class Waveform(QuantumInstruction):
-    # TODO: how to deal with the this not storing the actual pulse channel?
-    inst: Literal["Waveform"] = "Waveform"
-
-    @property
-    def channel(self):
-        return self.targets if isinstance(self.targets, str) else self.targets[0]
-
-
-class CustomPulse(Waveform):
-    """
-    Send a waveform with a custom shape down this particular channel.
-    """
-
-    # changed sample to numpy array: not sure of the consequences of this.
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-    inst: Literal["CustomPulse"] = "CustomPulse"
-    samples: np.ndarray
-    ignore_channel_scale: bool
-    sample_time: float
-
-    def __init__(
-        self,
-        targets: PulseChannel,
-        samples: np.ndarray,
-        ignore_channel_scale: bool = False,
-        sample_time=None,
-        **kwargs,
-    ):
-        # TODO: should we be storing properties of the pulse channel in the instruction?
-        # It is only used for calcualting the duration (which was previously done externally).
-        if not sample_time:
-            chan = targets
-            if isinstance(chan, list):
-                chan = chan[0]
-            if not isinstance(chan, PulseChannel):
-                raise (
-                    ValueError("Sample time cannot be determined without a pulse channel.")
-                )
-            sample_time = chan.sample_time
-
-        super().__init__(
-            targets=targets,
-            samples=samples,
-            ignore_channel_scale=ignore_channel_scale,
-            sample_time=sample_time,
-        )
-
-    @property
-    def duration(self):
-        return len(self.samples) * self.sample_time
-
-    def __repr__(self):
-        return f"custom pulse {self.channel},{self.duration}"
-
-    @field_validator("samples", mode="before")
-    @classmethod
-    def _samples_as_np_array(cls, val):
-        """
-        Used in deserialization.
-
-        Numpyas arrays of complex numbers are serialized as list of strings.
-        """
-        if not isinstance(val, np.ndarray):
-            return np.array([complex(v) for v in val])
-        return val
-
-    @field_serializer("samples", when_used="json")
-    def _np_array_to_list(self, samples):
-        return samples.tolist()
-
-
-class Pulse(Waveform):
-    """
-    Send a pulse down this particular channel.
-    """
-
-    # TODO: The amount of variables for different pulses here is overwhelming,
-    # and won't do us any favours for serialization. How we deal with different
-    # pulses is an open question, but a first step could be to standarize notation
-    # between different pulse shapes.
-    inst: Literal["Pulse"] = "Pulse"
-    shape: PulseShapeType
-    width: Union[Variable, float] = 0.0
-    amp: Union[Variable, float] = 1.0
-    phase: Union[Variable, float] = 0.0
-    drag: Union[Variable, float] = 0.0
-    rise: Union[Variable, float] = 0.0
-    amp_setup: Union[Variable, float] = 0.0
-    scale_factor: Union[Variable, float] = 1.0
-    zero_at_edges: bool = False
-    beta: Union[Variable, float] = 0.0
-    frequency: Union[Variable, float] = 0.0
-    internal_phase: Union[Variable, float] = 0.0
-    std_dev: Union[Variable, float] = 0.0
-    square_width: Union[Variable, float] = 0.0
-    ignore_channel_scale: bool = False
-
-    @property
-    def duration(self):
-        return self.width
-
-    def __repr__(self):
-        return (
-            f"pulse {self.channel},{self.shape.value},{self.amp},"
-            f"{self.phase},{self.width},{self.drag},{self.rise}"
-        )
-
-
-class MeasurePulse(Pulse):
-    inst: Literal["MeasurePulse"] = "MeasurePulse"
-
-
-class DrivePulse(Pulse):
-    inst: Literal["DrivePulse"] = "DrivePulse"
-
-
-class SecondStatePulse(Pulse):
-    inst: Literal["SecondStatePulse"] = "SecondStatePulse"
-
-
-class CrossResonancePulse(Pulse):
-    inst: Literal["CrossResonancePulse"] = "CrossResonancePulse"
-
-
-class CrossResonanceCancelPulse(Pulse):
-    inst: Literal["CrossResonanceCancelPulse"] = "CrossResonanceCancelPulse"
-
-
-# previously also inheritted QuantumComponent: I do not know why...
-# might need adding back later
-class Acquire(QuantumInstruction):
-    inst: Literal["Acquire"] = "Acquire"
-    suffix_incrementor: int = 0
-    time: Union[float, Variable] = 1e-6
-    mode: AcquireMode = (AcquireMode.RAW,)
-    output_variable: str = None
-    delay: Optional[float] = None
-    # was previously union of pulse and custom pulse: can this be extended to waveform?
-    filter: Optional[Waveform] = None
-
-    def __init__(
-        self,
-        targets: PulseChannel,
-        time: Union[float, Variable] = None,
-        mode: AcquireMode = None,
-        output_variable: Optional[str] = None,
-        existing_names: Set[str] = None,
-        delay: Optional[float] = None,
-        filter: Optional[Waveform] = None,
-        suffix_incrementor: int = 0,
-        **kwargs,
-    ):
-        # TODO: can't delegate output to validator as it requires "existing_names".
-        # Figure out the best way to do this (might just be as simple as replacing
-        # generate_name with something simplier...)
-        super().__init__(
-            targets=targets,
-            time=time or 1e-6,
-            mode=mode or AcquireMode.RAW,
-            delay=delay,
-            output_variable=output_variable
-            or self.generate_name(existing_names, targets.full_id()),
-            filter=filter,
-            suffix_incrementor=suffix_incrementor,
-        )
-
-    def generate_name(self, existing_names=None, channel=None):
-        return build_generated_name(existing_names, channel if channel else self.channel)
-
-    @property
-    def duration(self):
-        return self.time
-
-    @property
-    def channel(self):
-        return self.targets
-
-    def __repr__(self):
-        out_var = f"->{self.output_variable}" if self.output_variable else ""
-        mode = f",{self.mode.value}" if self.mode is not None else ""
-        return f"acquire {self.channel},{self.time}{mode}{out_var}"
-
-    @field_validator("filter", mode="before")
-    @classmethod
-    def _validate_filter(cls, filter, info: ValidationInfo):
-        if filter == None:
-            return filter
-
-        time = info.data["time"]
-        if isinstance(time, Variable):
-            get_default_logger.warning(f"Filter cannot be applied when time is variable.")
-            return None
-
-        if not isinstance(filter, Waveform):
-            raise ValueError("Filter must be a Waveform.")
-
-        if filter.duration < 0:
-            raise ValueError(
-                f"Filter duration {filter.duration} cannot be less than or equal to zero."
-            )
-
-        if not np.isclose(filter.duration, time, atol=1e-9):
-            raise ValueError(
-                f"Filter duration '{filter.duration}' must be equal to Acquire "
-                f"duration '{time}'."
-            )
-
-        return filter
-
-
-# TODO: Used to be a quantum instruction, determine if this needs to be
-class PostProcessing(Instruction):
-    """
-    States what post-processing should happen after data has been acquired. This can
-    happen in the FPGA's or a software post-process.
-    """
-
-    inst: Literal["PostProcessing"] = "PostProcessing"
-    acquire: Acquire
-    process: PostProcessType
-    axes: List[ProcessAxis] = []
-    args: List[Any] = []
-    result_needed: bool = False
-
-    @property
-    def output_variable(self):
-        return self.acquire.output_variable
-
-    def __repr__(self):
-        axis = ",".join([axi.value for axi in self.axes])
-        args = f",{','.join(str(arg) for arg in self.args)}" if len(self.args) > 0 else ","
-        output_var = f"->{self.output_variable}" if self.output_variable is not None else ""
-        return (
-            f"{self.process.value} {self.acquire.output_variable}{args}{axis}{output_var}"
-        )
-
-    @field_validator("axes", mode="before")
-    @classmethod
-    def _axes_as_list(cls, axes):
-        if axes:
-            return axes if isinstance(axes, List) else [axes]
-        return []
-
-
 class Reset(QuantumInstruction):
     """Resets this qubit to its starting state."""
 
@@ -821,135 +569,3 @@ class InstructionBlock(BaseModel):
             raise ValueError(f"Invalid {label} for {type(self)}: {invalid_items_str}")
 
         return items
-
-
-class MeasureData(BaseModel):
-    mode: AcquireMode
-    output_variable: Optional[str] = None
-    measure: MeasurePulse
-    acquire: Acquire
-    duration: float
-    # TODO: think of a better solution than storing here
-    targets: List[str] = None
-
-
-class MeasureBlock(InstructionBlock):
-    """Groups multiple qubit measurements together."""
-
-    inst: Literal["MeasureBlock"] = "MeasureBlock"
-    target_dict: Dict[str, MeasureData] = {}
-    existing_names: Optional[Set[str]] = set()
-    duration: float = 0.0
-
-    @property
-    def instructions(self):
-        instructions = [Synchronize(self.pulse_channel_targets)]
-        for val in self.target_dict.values():
-            instructions.extend([val.measure, val.acquire])
-        instructions.append(Synchronize(self.pulse_channel_targets))
-        return instructions
-
-    @property
-    def targets(self):
-        return list(self.target_dict.keys())
-
-    @property
-    def pulse_channel_targets(self):
-        targets = []
-        for target in self.target_dict.values():
-            targets.extend(target.targets)
-        return targets
-
-    @staticmethod
-    def create_block(
-        qubit: Union[Qubit, List[Qubit]],
-        mode: AcquireMode,
-        output_variables: str = None,
-        existing_names=set(),
-    ):
-        # set as a seperate static method as overwriting __init__ would
-        # make this unserialisable...
-        measure_block = MeasureBlock(existing_names=existing_names)
-        measure_block.add_measurements(qubit, mode, output_variables)
-        return measure_block
-
-    def add_measurements(
-        self,
-        targets: Union[Qubit, List[Qubit]],
-        mode: AcquireMode,
-        output_variables: Union[str, List[str]] = None,
-        existing_names: Set[str] = None,
-    ):
-        targets = self._validate_types(targets, (Qubit))
-        if len((duplicates := [t for t in targets if t.full_id() in self.targets])) > 0:
-            raise ValueError(
-                "Target can only be measured once in a 'MeasureBlock'. "
-                f"Duplicates: {duplicates}"
-            )
-        if not isinstance(output_variables, list):
-            output_variables = [] if output_variables is None else [output_variables]
-        if (num_out_vars := len(output_variables)) == 0:
-            output_variables = [None] * len(targets)
-        elif num_out_vars != len(targets):
-            raise ValueError(
-                f"Unsupported number of `output_variables`: {num_out_vars}, "
-                f"must be `None` or match numer of targets: {len(targets)}."
-            )
-        for target, output_variable in zip(targets, output_variables):
-            meas, acq = self._generate_measure_acquire(
-                target, mode, output_variable, existing_names
-            )
-            duration = max(meas.duration, acq.delay + acq.duration)
-            self.target_dict[target.full_id()] = MeasureData(
-                mode=mode,
-                output_variable=output_variable,
-                measure=meas,
-                acquire=acq,
-                duration=duration,
-                targets=[pc.full_id() for pc in target.pulse_channels.values()],
-            )
-            self.duration = max(self.duration, duration)
-
-    def _generate_measure_acquire(self, qubit, mode, output_variable, existing_names):
-        measure_channel = qubit.get_measure_channel()
-        acquire_channel = qubit.get_acquire_channel()
-        existing_names = existing_names or self.existing_names
-        weights = (
-            qubit.measure_acquire.get("weights", None)
-            if qubit.measure_acquire.get("use_weights", False)
-            else None
-        )
-
-        measure_instruction = MeasurePulse(measure_channel, **qubit.pulse_measure)
-        acquire_instruction = Acquire(
-            acquire_channel,
-            (
-                qubit.pulse_measure["width"]
-                if qubit.measure_acquire["sync"]
-                else qubit.measure_acquire["width"]
-            ),
-            mode,
-            output_variable,
-            existing_names,
-            qubit.measure_acquire["delay"],
-            weights,
-        )
-
-        return [
-            measure_instruction,
-            acquire_instruction,
-        ]
-
-    def get_acquires(self, targets: Union[Qubit, List[Qubit]]):
-        if not isinstance(targets, list):
-            targets = [targets]
-        targets = [t.full_id() if isinstance(t, Qubit) else t for t in targets]
-        return [self.target_dict[qt].acquire for qt in targets]
-
-    def __repr__(self):
-        target_strings = []
-        for q, d in self.target_dict.items():
-            out_var = f"->{d.output_variable}" if d.output_variable is not None else ""
-            mode = f":{d.mode.value}" if d.mode is not None else ""
-            target_strings.append(f"{q}{mode}{out_var}")
-        return f"Measure {', '.join(target_strings)}"
