+++ tests/qat/ir/test_instructions.py
@@ -1,4 +1,3 @@
-import copy
 from itertools import chain, product
 
 import numpy as np
@@ -6,41 +5,23 @@ import pytest
 from compiler_config.config import InlineResultsProcessing
 from pydantic import ValidationError
 
-from qat.ir.instruction_list import InstructionList
 from qat.ir.instructions import (
-    Acquire,
-    AcquireMode,
     Assign,
-    CrossResonanceCancelPulse,
-    CrossResonancePulse,
-    CustomPulse,
     Delay,
-    DrivePulse,
-    EndSweep,
     FrequencyShift,
     Jump,
     Label,
-    MeasureBlock,
-    MeasurePulse,
     PhaseReset,
     PhaseShift,
-    PostProcessing,
-    PostProcessType,
-    Pulse,
-    PulseShapeType,
     QuantumInstruction,
     Repeat,
     Reset,
     ResultsProcessing,
     Return,
-    SecondStatePulse,
-    Sweep,
     Synchronize,
     Variable,
 )
 from qat.purr.backends.echo import get_default_echo_hardware
-from qat.purr.compiler import instructions as LegacyInstructions
-from qat.utils.ir_converter import IRConverter
 
 
 class TestVariable:
@@ -336,103 +317,6 @@ class TestSynchronizePhaseReset:
         assert inst.targets == inst2.targets
 
 
-class TestCustomPulse:
-
-    model = get_default_echo_hardware()
-
-    def gaussian_shape(self, points, amp):
-        xs = np.linspace(-1, +1, points)
-        return amp * np.exp(-0.5 * (xs**2))
-
-    @pytest.mark.parametrize(
-        ["points", "amp", "ignore"],
-        product([3, 57, 101], [1.0, 1.656e8, 1.5e4 - 2.1e3j], [True, False]),
-    )
-    def test_custom_pulse(self, points, amp, ignore):
-        chan = list(self.model.pulse_channels.values())[0]
-        inst = CustomPulse(
-            chan,
-            self.gaussian_shape(points, amp),
-            ignore,
-        )
-        assert inst.targets == chan.full_id()
-        assert inst.sample_time == chan.physical_channel.sample_time
-        assert inst.duration == chan.physical_channel.sample_time * points
-
-
-class TestPulse:
-
-    model = get_default_echo_hardware()
-
-    @pytest.mark.parametrize(
-        ["pulse", "duration", "shape"],
-        product(
-            [
-                Pulse,
-                MeasurePulse,
-                DrivePulse,
-                SecondStatePulse,
-                CrossResonancePulse,
-                CrossResonanceCancelPulse,
-            ],
-            [0.0, 8e-8, 8e-6],
-            PulseShapeType,
-        ),
-    )
-    def test_waveform(self, pulse, duration, shape):
-        chan = list(self.model.pulse_channels.values())[0]
-        inst = pulse(chan, shape=shape, width=duration)
-        assert inst.width == duration
-        assert inst.duration == duration
-        assert inst.shape == shape
-
-    def test_invalid_shape(self):
-        with pytest.raises(AttributeError):
-            chan = list(self.model.pulse_channels.values())[0]
-            Pulse(chan, shape=PulseShapeType.CIRCLE, width=8e-8)
-
-    @pytest.mark.parametrize("duration", [8e-8j, "pi"])
-    def test_invalid_duraiton(self, duration):
-        with pytest.raises(ValidationError):
-            chan = list(self.model.pulse_channels.values())[0]
-            Pulse(chan, shape=PulseShapeType.SQUARE, width=duration)
-
-
-class TestAcquire:
-    model = get_default_echo_hardware()
-
-    def test_initiate(self):
-        chan = self.model.get_qubit(0).get_acquire_channel()
-        inst = Acquire(chan)
-        assert inst.time == 1e-6
-        assert inst.targets == chan.full_id()
-
-    def test_filter(self):
-        chan = self.model.get_qubit(0).get_acquire_channel()
-        filter = Pulse(chan, shape=PulseShapeType.GAUSSIAN, width=1e-6)
-        inst = Acquire(chan, 1e-6, filter=filter)
-        assert inst.filter == filter
-
-    @pytest.mark.parametrize("time", [0, 5e-7, 1.01e-6, 2e-6])
-    def test_filter_validation(self, time):
-        chan = self.model.get_qubit(0).get_acquire_channel()
-        filter = Pulse(chan, shape=PulseShapeType.GAUSSIAN, width=time)
-        with pytest.raises(ValidationError):
-            inst = Acquire(chan, 1e-6, filter=filter)
-
-
-class TestPostProcessing:
-    model = get_default_echo_hardware()
-
-    @pytest.mark.parametrize("pp", PostProcessType)
-    def test_initiate(self, pp):
-        chan = self.model.get_qubit(0).get_acquire_channel()
-        acquire = Acquire(chan)
-        inst = PostProcessing(acquire=acquire, process=pp)
-        assert inst.process == pp
-        assert inst.acquire == acquire
-
-
 class TestReset:
     model = get_default_echo_hardware()
 
@@ -452,263 +336,3 @@ class TestReset:
     def test_wrong_target(self):
         with pytest.raises(ValueError):
             Reset(self.model.get_qubit(0).measure_device)
-
-
-class TestInstructionBlocks:
-    # Tests are extracted and modified from test_instructions.py
-
-    @pytest.mark.parametrize("mode", list(AcquireMode))
-    @pytest.mark.parametrize("num_qubits", [1, 3])
-    def test_create_simple_measure_block(self, num_qubits, mode):
-        hw = get_default_echo_hardware()
-        targets = hw.qubits[:num_qubits]
-
-        mb = MeasureBlock.create_block(targets, mode)
-        assert isinstance(mb, MeasureBlock)
-        assert mb.targets == [t.full_id() for t in targets]
-        assert mb.target_dict[targets[0].full_id()].mode == mode
-
-    @pytest.mark.parametrize("out_vars", [None, "c"])
-    @pytest.mark.parametrize("num_qubits", [1, 3])
-    def test_create_measure_block_with_output_variables(self, num_qubits, out_vars):
-        hw = get_default_echo_hardware()
-        targets = hw.qubits[:num_qubits]
-
-        if isinstance(out_vars, str):
-            out_vars = [f"{out_vars}[{i}]" for i in range(num_qubits)]
-
-        mb = MeasureBlock.create_block(
-            targets, AcquireMode.INTEGRATOR, output_variables=out_vars
-        )
-        expected = out_vars or [None] * num_qubits
-        assert [val.output_variable for val in mb.target_dict.values()] == expected
-
-    def test_add_to_measure_block(self):
-        hw = get_default_echo_hardware()
-        targets = [hw.qubits[0], hw.qubits[-1]]
-        modes = [AcquireMode.INTEGRATOR, AcquireMode.SCOPE]
-        out_vars = ["c[0]", "b[1]"]
-        mb = MeasureBlock.create_block(
-            targets[0],
-            modes[0],
-            output_variables=out_vars[:1],
-        )
-        assert mb.targets == [t.full_id() for t in hw.qubits[:1]]
-        mb.add_measurements(targets[1], modes[1], output_variables=out_vars[1])
-        assert mb.targets == [t.full_id() for t in targets]
-        assert [val.mode for val in mb.target_dict.values()] == modes
-        assert [val.output_variable for val in mb.target_dict.values()] == out_vars
-
-    def test_cannot_add_duplicate_to_measure_block(self):
-        hw = get_default_echo_hardware()
-        targets = [hw.qubits[0], hw.qubits[-1]]
-        out_vars = ["c[0]", "b[1]"]
-        mb = MeasureBlock.create_block(
-            targets,
-            AcquireMode.INTEGRATOR,
-            output_variables=out_vars,
-        )
-        assert mb.targets == [t.full_id() for t in targets]
-        with pytest.raises(ValueError):
-            mb.add_measurements(targets[1], AcquireMode.INTEGRATOR)
-
-    def test_measure_block_duration(self):
-        hw = get_default_echo_hardware()
-        target = hw.qubits[0]
-        mb = MeasureBlock.create_block([], AcquireMode.RAW)
-        assert mb.duration == 0.0
-        mb.add_measurements(target, AcquireMode.INTEGRATOR)
-        acq = mb.get_acquires(target)[0]
-        assert mb.duration > 0
-        assert mb.duration == pytest.approx(acq.delay + acq.duration)
-        mb.duration = 1
-        assert mb.duration == 1
-
-    def test_get_acquires(self):
-        hw = get_default_echo_hardware()
-        mb = MeasureBlock.create_block(hw.qubits, AcquireMode.INTEGRATOR)
-        acquires = mb.get_acquires(hw.qubits)
-        assert all([isinstance(acq, Acquire) for acq in acquires])
-
-
-class TestInstructionList:
-
-    def all_instructions(self):
-        model = get_default_echo_hardware()
-        instruction_list = [
-            Repeat(repeat_count=1000, repetition_period=0.0005),
-            Assign(name="test", value=42),
-            Return(variables=["dave", "steve"]),
-            Sweep(),
-            EndSweep(),
-            Label.with_random_name(),
-            Jump(target="test"),
-            PhaseShift(model.get_qubit(0).get_drive_channel(), phase=np.pi),
-            FrequencyShift(model.get_qubit(1).get_measure_channel(), frequency=7e8),
-            Delay(model.get_qubit(0).get_drive_channel(), time=1e-8),
-            Synchronize(model.qubits),
-            MeasurePulse(
-                model.get_qubit(1).get_measure_channel(),
-                shape=PulseShapeType.SQUARE,
-                width=Variable(name="length", var_type=float, value=1e-6),
-            ),
-            DrivePulse(
-                model.get_qubit(0).get_drive_channel(),
-                shape=PulseShapeType.GAUSSIAN,
-                width=8e-8,
-            ),
-            Acquire(
-                model.get_qubit(0).get_acquire_channel(),
-                time=1e-6,
-                mode=AcquireMode.INTEGRATOR,
-            ),
-            Reset(model.get_qubit(1)),
-            PhaseReset(model.get_qubit(0)),
-            MeasureBlock.create_block(model.get_qubit(1), AcquireMode.RAW),
-        ]
-        return InstructionList(instruction_list=instruction_list)
-
-    def test_serialize(self):
-        instructions = self.all_instructions()
-        blob = instructions.serialize()
-        new_instructions = instructions.deserialize(blob)
-        for i, instruction in enumerate(instructions.instruction_list):
-            assert instruction == new_instructions.instruction_list[i]
-
-
-def make_pydantic_instruction_list():
-    hw = get_default_echo_hardware()
-    instructions = [
-        Repeat(repeat_count=1000, repetition_period=5e-4),
-        Assign(name="test", value=5),
-        Return(variables=["test", "dave"]),
-        Label(name="test"),
-        Jump(target="test"),
-        # ResultsProcessing("test", PostProcessType.MEAN),
-        PhaseShift(hw.get_qubit(0).get_drive_channel(), phase=-0.123),
-        FrequencyShift(hw.get_qubit(0).get_drive_channel(), fequency=1e8),
-        Delay(hw.get_qubit(0).get_drive_channel(), time=8e-8),
-        Synchronize(hw.qubits[0:3]),
-        Synchronize(hw.qubits[0]),
-        Synchronize(hw.qubits[0].get_measure_channel()),
-        CustomPulse(hw.get_qubit(1).get_drive_channel(), np.ones(101)),
-        Pulse(hw.get_qubit(0).get_drive_channel(), shape=PulseShapeType.SQUARE, width=8e-8),
-        MeasurePulse(
-            hw.get_qubit(0).get_measure_channel(), shape=PulseShapeType.GAUSSIAN, width=8e-8
-        ),
-        DrivePulse(
-            hw.get_qubit(0).get_drive_channel(), shape=PulseShapeType.SECH, width=8e-8
-        ),
-        CrossResonancePulse(
-            hw.get_qubit(0).get_cross_resonance_channel(hw.get_qubit(1)),
-            shape=PulseShapeType.SQUARE,
-            width=8e-8,
-        ),
-        CrossResonanceCancelPulse(
-            hw.get_qubit(1).get_cross_resonance_cancellation_channel(hw.get_qubit(0)),
-            shape=PulseShapeType.SQUARE,
-            width=8e-8,
-        ),
-        Acquire(hw.get_qubit(0).get_acquire_channel()),
-        PostProcessing(
-            acquire=Acquire(hw.get_qubit(1).get_acquire_channel()),
-            process=PostProcessType.MEAN,
-        ),
-        Reset(hw.get_qubit(0).get_drive_channel()),
-        Reset(hw.get_qubit(0)),
-        Reset(hw.qubits[0:2]),
-        PhaseReset(hw.get_qubit(0).get_drive_channel()),
-        PhaseReset(hw.get_qubit(0)),
-        PhaseReset(hw.qubits[0:2]),
-    ]
-
-    return hw, instructions
-
-
-def make_legacy_instruction_list(InstructionTypes):
-    hw = get_default_echo_hardware()
-    instructions = [
-        InstructionTypes.Repeat(1000, 5e-4),
-        InstructionTypes.Assign("test", 5),
-        InstructionTypes.Return(["test", "dave"]),
-        InstructionTypes.Label("test"),
-        InstructionTypes.Jump("test"),
-        # InstructionTypes.ResultsProcessing("test", PostProcessType.MEAN),
-        InstructionTypes.PhaseShift(hw.get_qubit(0).get_drive_channel(), -0.123),
-        InstructionTypes.FrequencyShift(hw.get_qubit(0).get_drive_channel(), 1e8),
-        InstructionTypes.Delay(hw.get_qubit(0).get_drive_channel(), 8e-8),
-        InstructionTypes.Synchronize(hw.qubits[0:3]),
-        InstructionTypes.Synchronize(hw.qubits[0]),
-        InstructionTypes.Synchronize(hw.qubits[0].get_measure_channel()),
-        InstructionTypes.CustomPulse(hw.get_qubit(1).get_drive_channel(), np.ones(101)),
-        InstructionTypes.Pulse(
-            hw.get_qubit(0).get_drive_channel(), PulseShapeType.SQUARE, 8e-8
-        ),
-        InstructionTypes.MeasurePulse(
-            hw.get_qubit(0).get_measure_channel(), PulseShapeType.GAUSSIAN, 8e-8
-        ),
-        InstructionTypes.DrivePulse(
-            hw.get_qubit(0).get_drive_channel(), PulseShapeType.SECH, 8e-8
-        ),
-        InstructionTypes.CrossResonancePulse(
-            hw.get_qubit(0).get_cross_resonance_channel(hw.get_qubit(1)),
-            PulseShapeType.SQUARE,
-            8e-8,
-        ),
-        InstructionTypes.CrossResonanceCancelPulse(
-            hw.get_qubit(1).get_cross_resonance_cancellation_channel(hw.get_qubit(0)),
-            PulseShapeType.SQUARE,
-            8e-8,
-        ),
-        InstructionTypes.Acquire(hw.get_qubit(0).get_acquire_channel()),
-        InstructionTypes.PostProcessing(
-            InstructionTypes.Acquire(hw.get_qubit(1).get_acquire_channel()),
-            PostProcessType.MEAN,
-        ),
-        InstructionTypes.Reset(hw.get_qubit(0).get_drive_channel()),
-        InstructionTypes.Reset(hw.get_qubit(0)),
-        InstructionTypes.Reset(hw.qubits[0:2]),
-        InstructionTypes.PhaseReset(hw.get_qubit(0).get_drive_channel()),
-        InstructionTypes.PhaseReset(hw.get_qubit(0)),
-        InstructionTypes.PhaseReset(hw.qubits[0:2]),
-    ]
-
-    return hw, instructions
-
-
-def eq_instructions(inst1, inst2):
-    dict1 = copy.copy(inst1.__dict__)
-    dict2 = copy.copy(inst2.__dict__)
-    if dict1.keys() != dict2.keys():
-        return False
-    for attr in dict1.keys():
-        if type(dict1[attr]) != type(dict2[attr]):
-            return False
-        if isinstance(dict1[attr], list):
-            for itm in dict1[attr]:
-                if not itm in dict2[attr]:
-                    return False
-    return True
-
-
-hw, legacy_instructions = make_legacy_instruction_list(LegacyInstructions)
-hw, pydantic_instructions = make_pydantic_instruction_list()
-
-
-class TestConversion:
-
-    @pytest.mark.parametrize("instruction", legacy_instructions)
-    @pytest.mark.parametrize("hw", [hw])
-    def test_legacy_to_pydantic(self, instruction, hw, monkeypatch):
-        monkeypatch.setattr(LegacyInstructions.Instruction, "__eq__", eq_instructions)
-        pydantic_inst = IRConverter()._legacy_to_pydantic_instruction(instruction)
-        legacy_inst = IRConverter(hw)._pydantic_to_legacy_instruction(pydantic_inst)
-        assert legacy_inst == instruction
-
-    @pytest.mark.parametrize("instruction", pydantic_instructions)
-    @pytest.mark.parametrize("hw", [hw])
-    def test_pydantic_to_legacy(self, instruction, hw, monkeypatch):
-        legacy_inst = IRConverter(hw)._pydantic_to_legacy_instruction(instruction)
-        pydantic_inst = IRConverter()._legacy_to_pydantic_instruction(legacy_inst)
-        assert pydantic_inst == instruction
-        assert instruction.inst == pydantic_inst.inst
